requirements.txt
apscheduler
beautifulsoup4
pyyaml
requests
retry
selenium
tqdm
